change the type of `AssignStack::num_best_assign` to f64 to let it decay
	modified:   src/assign/mod.rs
	modified:   src/assign/propagate.rs
	modified:   src/assign/stack.rs
	modified:   src/config.rs
	modified:   src/solver/conflict.rs
	modified:   src/state.rs

+++ src/assign/mod.rs
@@ -135,7 +135,7 @@ pub struct AssignStack {
     use_rephase: bool,
     best_assign: bool,
     build_best_at: usize,
-    num_best_assign: usize,
+    num_best_assign: f64,
 
     //
     //## Statistics
